Move generateGame function to separate file
